+++ Tests/Runtime/RateLimitTests.cs
@@ -1,5 +1,4 @@
-﻿using Backtrace.Unity;
-using Backtrace.Unity.Model;
+﻿using Backtrace.Unity.Model;
 using NUnit.Framework;
 using System;
 using System.Collections;
@@ -8,7 +7,7 @@ using UnityEngine.TestTools;
 
 namespace Backtrace.Unity.Tests.Runtime
 {
-    public class RateLimitTests: BacktraceBaseTest
+    public class RateLimitTests : BacktraceBaseTest
     {
         private const int CLIENT_RATE_LIMIT = 3;
 
@@ -20,10 +19,11 @@ namespace Backtrace.Unity.Tests.Runtime
             BacktraceClient.SetClientReportLimit(CLIENT_RATE_LIMIT);
             AfterSetup();
         }
-        [TestCase(5)]
-        [TestCase(10)]
-        [TestCase(20)]
-        public void TestReportLimit_ShouldntHitRateLimit_AllReportsShouldBeInBacktrace(int reportPerMin)
+
+        private static int[] _rateLimitTest = new int[] { 5, 10, 20 };
+
+        [UnityTest]
+        public IEnumerator TestReportLimit_ShouldntHitRateLimit_AllReportsShouldBeInBacktrace([ValueSource("_rateLimitTest")] int reportPerMin)
         {
             uint rateLimit = Convert.ToUInt32(reportPerMin);
             BacktraceClient.SetClientReportLimit(rateLimit);
@@ -42,6 +42,8 @@ namespace Backtrace.Unity.Tests.Runtime
             {
                 BacktraceClient.Send("test");
             }
+
+            yield return new WaitForEndOfFrame();
             Assert.AreEqual(maximumNumberOfRetries, rateLimit);
             Assert.AreEqual(0, skippedReports);
         }
@@ -67,6 +69,8 @@ namespace Backtrace.Unity.Tests.Runtime
             {
                 BacktraceClient.Send("test");
             }
+
+            yield return new WaitForEndOfFrame();
             Assert.AreEqual(totalNumberOfReports, maximumNumberOfRetries + skippedReports);
             Assert.AreEqual(maximumNumberOfRetries, CLIENT_RATE_LIMIT);
             Assert.AreEqual(totalNumberOfReports - CLIENT_RATE_LIMIT, skippedReports);
@@ -95,6 +99,8 @@ namespace Backtrace.Unity.Tests.Runtime
                 BacktraceClient.Send(new Exception("Exception"));
 
             }
+
+            yield return new WaitForEndOfFrame();
             Assert.AreEqual(totalNumberOfReports, maximumNumberOfRetries + skippedReports);
             Assert.AreEqual(maximumNumberOfRetries, CLIENT_RATE_LIMIT);
             Assert.AreEqual(totalNumberOfReports - CLIENT_RATE_LIMIT, skippedReports);
@@ -124,34 +130,8 @@ namespace Backtrace.Unity.Tests.Runtime
                 var report = new BacktraceReport(new Exception("Exception"));
                 BacktraceClient.Send(report);
             }
-            Assert.AreEqual(totalNumberOfReports, maximumNumberOfRetries + skippedReports);
-            Assert.AreEqual(maximumNumberOfRetries, CLIENT_RATE_LIMIT);
-            Assert.AreEqual(totalNumberOfReports - CLIENT_RATE_LIMIT, skippedReports);
-            yield return null;
-        }
 
-        [UnityTest]
-        public IEnumerator TestReportLimit_InvalidReportNumber_IgnoreAdditionalReports()
-        {
-            BacktraceClient.SetClientReportLimit(CLIENT_RATE_LIMIT);
-            int totalNumberOfReports = 5;
-            int maximumNumberOfRetries = 0;
-            BacktraceClient.RequestHandler = (string url, BacktraceData data) =>
-             {
-                 maximumNumberOfRetries++;
-                 return new BacktraceResult();
-             };
-            int skippedReports = 0;
-            BacktraceClient.OnClientReportLimitReached = (BacktraceReport report) =>
-            {
-                skippedReports++;
-            };
-
-            for (int i = 0; i < totalNumberOfReports; i++)
-            {
-                BacktraceClient.Send("test");
-
-            }
+            yield return new WaitForEndOfFrame();
             Assert.AreEqual(totalNumberOfReports, maximumNumberOfRetries + skippedReports);
             Assert.AreEqual(maximumNumberOfRetries, CLIENT_RATE_LIMIT);
             Assert.AreEqual(totalNumberOfReports - CLIENT_RATE_LIMIT, skippedReports);
@@ -172,6 +152,8 @@ namespace Backtrace.Unity.Tests.Runtime
             {
                 BacktraceClient.Send("test");
             }
+
+            yield return new WaitForEndOfFrame();
             Assert.AreEqual(2, maximumNumberOfRetries);
             yield return null;
         }