+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Backtrace Unity Release Notes
 
+## Version 3.0.2
+- `BacktraceDatabase` now allows to send object via `Send` method. This method will try to send all objects from database - with all respects to the Database settings.
+- `BacktraceClient` will try to reuse database JSON - if Backtrace client has acccess to Backtrace database,
+- `BacktraceDatabase` `AutoSend` property improvements - now we will discard any calculations in `Update` method,
+- `BacktraceClient` by default will generate configuraiton file with client rate limit equal to 50.
+
 ## Version 3.0.1
 - The `BacktraceDatabase` class will now create database directory before final database validation. Previously, when directory didn't exist, BacktraceDatabase was disabled.
 - The `BacktraceDatabase` field now allows users to pass interpolated string in Database options. Developer can use `${Application.dataPath}` or `${Application.persistentDataPath}` to set path to database.